# openapi.yaml
openapi: 3.0.3
info:
  title: FinApp API
  description: API for financial application
  version: 1.3.0
  contact:
    name: API Support
    email: support@finapp.com
servers:
  - url: http://localhost:8080/api/v1
    description: Local development server
  - url: https://api.finapp.com/api/v1
    description: Production server
paths:
...
  /categories:
    get:
      operationId: getCategories
      tags: [Categories]
      summary: Get categories list
      description: |
        Получает список всех категорий пользователя.
        Возвращает иерархическую структуру с подкатегориями.
      responses:
        '200':
          description: Categories list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CategoryDto'
        '401':
          description: Unauthorized access
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
    post:
      operationId: createCategory
      tags: [Categories]
      summary: Create new category
      description: |
        Создает новую категорию для классификации транзакций.
        Поддерживает создание подкатегорий и настройку иконок.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCategoryRequest'
      responses:
        '201':
          description: Category created
          headers:
            Location:
              schema:
                type: string
              description: "URL созданной категории"
              example: "/api/v1/categories/456e7890-e89b-12d3-a456-426614174001"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CategoryDto'
        '400':
          description: Invalid request data
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '409':
          description: Category name already exists
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
  /categories/{id}:
    get:
      operationId: getCategoryById
      tags: [Categories]
      summary: Get category by ID
      description: |
        Получает детальную информацию о категории по её идентификатору.
      parameters:
        - name: id
          in: path
          description: Category ID
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Category details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CategoryDto'
        '404':
          description: Category not found
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '401':
          description: Unauthorized
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
    put:
      operationId: updateCategory
      tags: [Categories]
      summary: Update category
      description: |
        Полностью обновляет данные категории.
      parameters:
        - name: id
          in: path
          description: Category ID
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateCategoryRequest'
      responses:
        '200':
          description: Category updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CategoryDto'
        '404':
          description: Category not found
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '400':
          description: Invalid request data
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '401':
          description: Unauthorized
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
    delete:
      operationId: deleteCategory
      tags: [Categories]
      summary: Delete category
      description: |
        Удаляет категорию. Если в категории есть транзакции, они будут перенесены в категорию "Без категории".
      parameters:
        - name: id
          in: path
          description: Category ID
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Category deleted successfully
        '404':
          description: Category not found
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '401':
          description: Unauthorized
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'

  # Receipt endpoints
...
components:
  schemas:
...
    CategoryDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: "ID категории"
        name:
          type: string
          description: "Название категории"
        type:
          $ref: '#/components/schemas/CategoryType'
        parentId:
          type: string
          format: uuid
          nullable: true
          description: "ID родительской категории"
        icon:
          type: string
          nullable: true
          description: "Иконка категории"
          example: "solar:wallet-money-line-duotone"
        iconColor:
          type: string
          nullable: true
          description: "Цвет иконки в HEX-формате"
          example: "#10B981"
        children:
          type: array
          items:
            $ref: '#/components/schemas/CategoryDto'
          description: "Список вложенных подкатегорий"
    CreateCategoryRequest:
      type: object
      required: [name, type]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 50
          description: "Название категории"
        type:
          $ref: '#/components/schemas/CategoryType'
        parentId:
          type: string
          format: uuid
          nullable: true
          description: "ID родительской категории"
        icon:
          type: string
          nullable: true
          description: "Иконка категории"
        iconColor:
          type: string
          nullable: true
          description: "Цвет иконки в HEX-формате"
    CategoryType:
      type: string
      enum: [INCOME, EXPENSE]
      description: "Тип категории"

    # Common DTOs
    ProblemDetails:
      type: object
      required: [type, title, status]
      properties:
        type:
          type: string
          format: uri
          description: "URI идентифицирующий тип проблемы"
          example: "https://api.finapp.com/problems/validation-error"
        title:
          type: string
          description: "Краткое описание типа проблемы"
          example: "Validation Error"
        status:
          type: integer
          description: "HTTP статус код"
          example: 400
        detail:
          type: string
          description: "Детальное описание проблемы"
          example: "Поле 'amount' должно быть положительным числом"
        instance:
          type: string
          format: uri
          description: "URI идентифицирующий конкретный случай проблемы"
          example: "/api/v1/transactions/123"
        errors:
          type: array
          description: "Детали ошибок валидации"
          items:
            type: object
            properties:
              field:
                type: string
                description: "Поле с ошибкой"
                example: "amount"
              message:
                type: string
                description: "Сообщение об ошибке"
                example: "должно быть положительным числом"
              code:
                type: string
                description: "Код ошибки"
                example: "POSITIVE_NUMBER_REQUIRED"

    # User DTOs
...
    UpdateCategoryRequest:
      type: object
      required: [name, type]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 100
          description: "Название категории"
        type:
          $ref: '#/components/schemas/CategoryType'
        icon:
          type: string
          nullable: true
          description: "Иконка категории"
        color:
          type: string
          nullable: true
          pattern: '^#([A-Fa-f0-9]{6}|[A-Fa-f0-9]{3})$'
          description: "Цвет категории в HEX формате"
        parentId:
          type: string
          format: uuid
          nullable: true
          description: "ID родительской категории"
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - BearerAuth: []